requirements.txt
# Core dependencies
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1

# PDF processing
PyMuPDF==1.23.8
pymupdf4llm==0.0.5

# Text processing and embeddings
sentence-transformers==2.2.2
chromadb==0.4.18
faiss-cpu==1.7.4

# Data processing
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2

# Web interface
streamlit==1.28.1
gradio==4.7.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1
loguru==0.7.2

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0
